# shema-christianzip/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StackGroup {
  category: string;
  items: string[];
}

export interface Project {
  id: number;
  slug: string;
  title: string;
  tagline: string;
  description: string;
  challenge: string;
  solution: string;
  outcome: string;
  tags: string[];
  image: string;
  link?: string | null;
  live: boolean;
  stack: StackGroup[];
  createdAt?: string | null;
  updatedAt?: string | null;
}

export interface ContentSection {
  type: string;
  text?: string;
  lang?: string | null;
  items?: string[] | null;
}

export interface Post {
  id: number;
  slug: string;
  title: string;
  excerpt: string;
  date: string;
  readTime: string;
  tag: string;
  tagColor: string;
  content: ContentSection[];
  createdAt?: string | null;
  updatedAt?: string | null;
}