# shema-christianzip/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Portfolio projects
  - name: posts
    description: Blog posts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: Array of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{slug}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all blog posts
      responses:
        "200":
          description: Array of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /posts/{slug}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    StackGroup:
      type: object
      properties:
        category:
          type: string
        items:
          type: array
          items:
            type: string
      required:
        - category
        - items

    Project:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        tagline:
          type: string
        description:
          type: string
        challenge:
          type: string
        solution:
          type: string
        outcome:
          type: string
        tags:
          type: array
          items:
            type: string
        image:
          type: string
        link:
          type: string
          nullable: true
        live:
          type: boolean
        stack:
          type: array
          items:
            $ref: "#/components/schemas/StackGroup"
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - slug
        - title
        - tagline
        - description
        - challenge
        - solution
        - outcome
        - tags
        - image
        - live
        - stack

    ContentSection:
      type: object
      properties:
        type:
          type: string
        text:
          type: string
        lang:
          type: string
          nullable: true
        items:
          type: array
          nullable: true
          items:
            type: string
      required:
        - type

    Post:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        date:
          type: string
        readTime:
          type: string
        tag:
          type: string
        tagColor:
          type: string
        content:
          type: array
          items:
            $ref: "#/components/schemas/ContentSection"
        createdAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - slug
        - title
        - excerpt
        - date
        - readTime
        - tag
        - tagColor
        - content