# shema-christianzip/lib/api-zod/src/generated/types/contentSection.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentSection {
  type: string;
  text?: string;
  lang?: string | null;
  items?: string[] | null;
}