# shema-christianzip/lib/api-zod/src/generated/types/project.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StackGroup } from './stackGroup';

export interface Project {
  id: number;
  slug: string;
  title: string;
  tagline: string;
  description: string;
  challenge: string;
  solution: string;
  outcome: string;
  tags: string[];
  image: string;
  link?: string | null;
  live: boolean;
  stack: StackGroup[];
  createdAt?: string | null;
  updatedAt?: string | null;
}